feat: migrate feeds worker to next api
