+++ apps/api/utils/feeds/providers/hey/heyFeed.ts
@@ -1,18 +1,12 @@
-import type { Env } from '../../types';
 import heyMostInteracted from './algorithms/heyMostInteracted';
 import heyMostViewed from './algorithms/heyMostViewed';
 
-const heyFeed = async (
-  strategy: string,
-  limit: number,
-  offset: number,
-  env: Env
-) => {
+const heyFeed = async (strategy: string, limit: number, offset: number) => {
   switch (strategy) {
     case 'mostviewed':
-      return await heyMostViewed(limit, offset, env);
+      return await heyMostViewed(limit, offset);
     case 'mostinteracted':
-      return await heyMostInteracted(limit, offset, env);
+      return await heyMostInteracted(limit, offset);
     default:
       return [];
   }
