+++ package.json
@@ -16,7 +16,7 @@
     "prettier:fix": "turbo run prettier:fix --parallel",
     "start": "turbo run start --parallel",
     "test:dev": "turbo run test:dev",
-    "test:e2e": "start-server-and-test start '8083|8086|8089' test:dev",
+    "test:e2e": "start-server-and-test start '8083|8086' test:dev",
     "typecheck": "turbo run typecheck --parallel",
     "typecheck:affected": "turbo run typecheck --parallel --filter=[origin/main]",
     "workers:deploy": "turbo run worker:deploy --parallel"
