--- packages/workers/feeds/.dev.vars.example
@@ -1,1 +1,0 @@
-CLICKHOUSE_PASSWORD=""
