--- packages/workers/feeds/.eslintrc.js
@@ -1,6 +1,0 @@
-module.exports = {
-  extends: [require.resolve('@hey/config/eslint/base.js')],
-  rules: {
-    'import/no-anonymous-default-export': 'off'
-  }
-};
