--- packages/workers/feeds/README.md
@@ -1,1 +1,0 @@
-# Feeds worker
