--- packages/workers/feeds/package.json
@@ -1,31 +1,0 @@
-{
-  "name": "@workers/feeds",
-  "version": "0.0.0",
-  "private": true,
-  "license": "AGPL-3.0",
-  "scripts": {
-    "dev": "wrangler dev --port 8089",
-    "lint": "eslint . --ext .ts",
-    "lint:fix": "eslint . --fix --ext .ts",
-    "prettier": "prettier --check \"**/*.{js,ts,tsx,md}\"  --cache",
-    "prettier:fix": "prettier --write \"**/*.{js,ts,tsx,md}\"  --cache",
-    "start": "pnpm dev",
-    "test:dev": "vitest --run",
-    "typecheck": "tsc --pretty",
-    "worker:deploy": "wrangler deploy --var RELEASE:\"$(git rev-parse HEAD)\""
-  },
-  "dependencies": {
-    "@hey/clickhouse": "workspace:*",
-    "@hey/data": "workspace:*",
-    "@hey/lib": "workspace:*",
-    "itty-router": "^4.0.23",
-    "urlcat": "^3.1.0"
-  },
-  "devDependencies": {
-    "@cloudflare/workers-types": "^4.20231025.0",
-    "@hey/config": "workspace:*",
-    "typescript": "^5.2.2",
-    "vitest": "^0.34.5",
-    "wrangler": "^3.15.0"
-  }
-}
