--- packages/workers/feeds/src/constants.ts
@@ -1,1 +1,0 @@
-export const TEST_URL = 'http://127.0.0.1:8089';
