--- packages/workers/feeds/src/providers/hey/README.md
@@ -1,4 +1,0 @@
-# Hey Algorithms
-
-- [MostViewed](./docs/MostViewed.md)
-- [MostInteracted](./docs/MostInteracted.md)
