--- packages/workers/feeds/src/providers/hey/docs/MostInteracted.md
@@ -1,17 +1,0 @@
-```ts
-const interactionAndWeights = {
-  [PUBLICATION.COLLECT_MODULE.COLLECT]: 10,
-  [PUBLICATION.MIRROR]: 8,
-  [PUBLICATION.SHARE]: 6,
-  [PUBLICATION.LIKE]: 5,
-  [PUBLICATION.ATTACHMENT.AUDIO.PLAY]: 4,
-  [PUBLICATION.ATTACHMENT.IMAGE.OPEN]: 4,
-  [PUBLICATION.TOGGLE_BOOKMARK]: 3,
-  [PUBLICATION.OPEN_MIRRORS]: 2,
-  [PUBLICATION.OPEN_LIKES]: 2,
-  [PUBLICATION.OPEN_COLLECTORS]: 2,
-  [PUBLICATION.COPY_TEXT]: 1,
-  [PUBLICATION.TRANSLATE]: 1,
-  [PUBLICATION.CLICK_OEMBED]: 1
-};
-```
