--- packages/workers/feeds/src/providers/k3l/README.md
@@ -1,1 +1,0 @@
-# Karma3Labs Algorithm
