--- packages/workers/feeds/tsconfig.json
@@ -1,6 +1,0 @@
-{
-  "extends": "@hey/config/base.tsconfig.json",
-  "compilerOptions": {
-    "types": ["@cloudflare/workers-types"]
-  }
-}
