--- packages/workers/feeds/vitest.config.ts
@@ -1,5 +1,0 @@
-import { defineConfig } from 'vitest/config';
-
-export default defineConfig({
-  test: { globals: true, testTimeout: 30000 }
-});
