--- packages/workers/feeds/wrangler.toml
@@ -1,12 +1,0 @@
-name = "feeds"
-main = "src/index.ts"
-compatibility_date = "2023-10-25"
-keep_vars = true
-
-routes = [
-	{ pattern = "feeds.hey.xyz", custom_domain = true }
-]
-
-[env.production.vars]
-RELEASE = ""
-CLICKHOUSE_PASSWORD = ""
